Handle errors in MultiMetricReader
Previously the error would be lost and the reader would halt if one of the
readers in a MultiMetricReader returned an error. We think this is the cause
of `ops.users.count` not being reported occasionally.

Handle this by sending the error to the buffer (which will be surfaced by
`ReadMetric()`) and restarting `CopyMetrics()` in a loop. We test this by
sending an error to one of the readers before sending the real metrics,
checking that we get our synthetic error, and then checking that we get all
of the metrics as before.

We had to change the type of `buffers` in order to have access to the
unexported `events` field. This means that we no longer need to do a type
assertion for `WriteMetrics()`. A new slice, of the interface type
`MetricReader`, then had to be made for `NewMultiMetricReader()`. Errors are
collected as a slice to check that we only get one and of the right content.