Define rep tls properties at the top-level.
There were erroneously added as properties.diego.rep.tls.* in 18cda83f5,
whereas they are actually defined at the top level
(properties.tls.*)[1]. Because they only apply to rep, this moves them
into the properties for the specific job. Ideally we'd specify them as
properties for rep alone, but that's not supported until we've migrated
to the V2 Bosh manifest schema.

[1]http://bosh.io/jobs/rep?source=github.com/cloudfoundry/diego-release&version=1.26.0#p=tls