Enable mTLS communication to rep on the cells.
Following the doc[1], this adds server and client certs for rep.

Initially require_tls is set to false on the client end, and true on the
server end. This is because the clients get upgraded first in our
deployment. Once upgraded, they will try TLS first, and fall back to
unencrypted. As the cells are upgraded, these connections will then be
re-established using TLS. A follow-up deploy will then set require_tls
to true on the clients.

There seems to be some conflicting information in the diego docs around
client certs, and whether a single client cert should be used for all
connections to a service, or whether each client should have its own
client cert. I've opened an issue[2] to get this clarified. In the
meantime, I've chosen to follow the exsiting pattern, and what the diego
manifest generation scripts do.

[1]https://github.com/cloudfoundry/diego-release/blob/v1.26.0/docs/upgrading-secure-cell-rep-api.md
[2]https://github.com/cloudfoundry/diego-release/issues/365